Fix Long < Double operator error in NotificationWorker
Kotlin Long has no operator compareTo(Double); cast the
Double result to Long before the comparison.

+++ app/src/main/java/com/vrijgeld/worker/NotificationWorker.kt
@@ -74,7 +74,7 @@ class NotificationWorker(
                 val totalBalance = accDao.getActiveOnce()
                     .sumOf { txRepo.getAccountBalance(it.id) }
                 dueSoon.forEach { sub ->
-                    if (totalBalance < sub.estimatedAmount * 1.5) {
+                    if (totalBalance < (sub.estimatedAmount * 1.5).toLong()) {
                         val days = ((sub.nextExpectedDate - now) / MS_PER_DAY).toInt().coerceAtLeast(0)
                         notif.notifyBillDueLowBalance(sub.merchantName, sub.estimatedAmount, days)
                     }
